refactor: lib폴더 내 함수 utils로 확장 및 fullReset 함수 및 폴더명 변경
